move serialization classes into azure.core.serialization
