+++ sdk/core/Azure.Core/tests/ModelSerializationTests/CatReadOnlyProperty.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text.Json;
+using Azure.Core.Serialization;
 
 namespace Azure.Core.Tests.ModelSerializationTests
 {
