+++ sdk/core/Azure.Core/tests/ModelSerializationTests/VerifyModels.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using Azure.Core.Serialization;
 using NUnit.Framework;
 
 namespace Azure.Core.Tests.ModelSerializationTests
